requirements.txt
python-vlc
pytest
apscheduler
mock
requests-mock
lxml
requests